# api.yaml
openapi: 3.0.3
info:
  title: Hospital Management System API
  description: >
    REST API for a hospital management system supporting three roles —
    Admin, Doctor, and Patient — with JWT-based authentication, appointment
    scheduling, treatment records, AI-powered patient summaries, and a
    post-appointment chat feature.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Unauthorized

    # ─── Doctor ──────────────────────────────────────────────────────────────
    Doctor:
      type: object
      properties:
        doctor_id:
          type: integer
        doctor_username:
          type: string
        doctor_name:
          type: string
        doctor_email:
          type: string
          format: email
        doctor_type:
          type: string
        doctor_qualification:
          type: string
        doctor_experience:
          type: string
        doctor_description:
          type: string
        dep_id:
          type: integer
          nullable: true
...
    Patient:
      type: object
      properties:
        patient_id:
          type: integer
        patient_name:
          type: string
        patient_username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          type: string
        age:
          type: integer
        address:
          type: string
        is_blacklisted:
          type: boolean
...
    PatientUpdateRequest:
      type: object
      properties:
        patient_name:
          type: string
        patient_username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          type: string
        age:
          type: integer
        address:
          type: string

    # ─── Department ──────────────────────────────────────────────────────────
    Department:
      type: object
      properties:
        dep_id:
          type: integer
        dep_name:
          type: string
        dep_des:
          type: string
        no_docs_registered:
          type: integer
...
    Appointment:
      type: object
      properties:
        app_id:
          type: integer
        patient_id:
          type: integer
        doctor_id:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
          format: time
          example: "09:00"
        status:
          type: string
          enum: [BOOKED, COMPLETED, CANCELLED]
        visit_type:
          type: string
          enum: [Morning, Afternoon, Evening]
        token_no:
          type: integer
          nullable: true
          description: Queue position within the slot (patient dashboard only)

    # ─── Treatment ───────────────────────────────────────────────────────────
...
    Availability:
      type: object
      properties:
        availability_id:
          type: integer
        doctor_id:
          type: integer
        date:
          type: string
          format: date
        slot_name:
          type: string
          enum: [Morning, Afternoon, Evening]
        start_time:
          type: string
          example: "09:00"
        end_time:
          type: string
          example: "12:00"
...
paths:
...
  /api/patient/profile:
    put:
      tags: [Patient]
      summary: Update the authenticated patient's own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdateRequest'
      responses:
        "200":
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Patient'
        "400":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/patient/departments/{dep_id}:
    get:
      tags: [Patient]
      summary: View a department and its list of doctors
      parameters:
        - in: path
          name: dep_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department and doctors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          department:
                            $ref: '#/components/schemas/Department'
                          doctors:
                            type: array
                            items:
                              $ref: '#/components/schemas/Doctor'
  /api/patient/doctors/{doctor_id}:
    get:
      tags: [Patient]
      summary: View a doctor's profile, availability, and per-slot booking counts
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details with availability
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          doctor:
                            $ref: '#/components/schemas/Doctor'
                          availability:
                            type: array
                            items:
                              $ref: '#/components/schemas/Availability'
                          slot_bookings:
                            type: object
                            description: availability_id → current booking count
                            additionalProperties:
                              type: integer
                          slot_capacity:
                            type: integer
                            example: 10
  /api/patient/book/{availability_id}:
    post:
      tags: [Patient]
      summary: Book an appointment for a given availability slot
      parameters:
        - in: path
          name: availability_id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Appointment booked successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Appointment'
        "400":
          description: Slot is in the past
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Slot fully booked or patient already has an appointment at this time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/patient/appointments/{appointment_id}/cancel:
    post:
      tags: [Patient]
      summary: Cancel a patient's own BOOKED appointment
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "400":
          description: Appointment already completed or cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Appointment belongs to a different patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/patient/export/csv:
    post:
      tags: [Patient]
      summary: Queue a Celery task to email the patient their appointment history as CSV
      responses:
        "200":
          description: Export task queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          task_id:
                            type: string
        "400":
          description: No email address on file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ────────────────────────────── CHAT ────────────────────────────────────────
...